updated crawler script to get player seasonal stats and telemetry match data

+++ crawler.py
@@ -9,16 +9,19 @@ with open("config.json") as config_file:
     API_KEY = json.load(config_file)["API_KEY"]
 
 PLAYER_API_REQUEST_ENDPOINT = "https://api.pubg.com/shards/kakao/players?filter[playerNames]="
-MATCH_API_REQUEST_ENDPOINT = "https://api.pubg.com/shards/steam/matches/"
-BANNED_PLAYERS_DIRECTORY = "player_ban_list"
+MATCH_API_REQUEST_ENDPOINT = "https://api.pubg.com/shards/kakao/matches/"
+PLAYERS_ID_PATH = "./player_ban_list"
+PLAYERS_DATA_PATH = "./banned_players_data/"
 
 header = {
   "Authorization": API_KEY,
-  "Accept": "application/vnd.api+json"
+  "Accept": "application/vnd.api+json",
+  "Accept-Encoding": "gzip"
 }
+metadata = {}
 
 # retrieve all relavent match statistics for specified player
-def requestMatchStats(match_id, player_id, headers, match_url = "https://api.pubg.com/shards/steam/matches/"):
+def requestMatchStats(match_id, player_id, headers, match_url = "https://api.pubg.com/shards/kakao/matches/"):
     match_response = requests.get(match_url+match_id, headers=headers)
     print("fetched match data",match_id,"with response status:",match_response.status_code)
 
@@ -61,7 +64,7 @@ def requestMatchStats(match_id, player_id, headers, match_url = "https://api.pub
                         roster_id_list[i] = {"id":data["attributes"]["stats"]["playerId"],
                                                 "name":data["attributes"]["stats"]["name"]}
 
-        return {"responseStatus": match_response,
+        return {"responseStatus": match_response.status_code,
                 "matchId": match_id,
                 "matchAttributes": match_attributes,
                 "playerId": player_id,
@@ -70,11 +73,28 @@ def requestMatchStats(match_id, player_id, headers, match_url = "https://api.pub
                 "won": won,
                 "rosterIdList": roster_id_list,
                 "telemetryDataURL": telemetry_URL}
-    return {"response_status":match_response}
+    return {"responseStatus":match_response.status_code}
 
-def writeAllMatchStatsToCsv(filename, player_id, match_list):
+def downloadAndWriteMatchTelemetryData(filename, telemetry_URL, headers, overwrite = False):
+    if (not overwrite and os.path.isfile(filename)):
+        print("filename",filename,"already exists. Aborting...")
+        return
+    response = requests.get(telemetry_URL, headers=headers)
+    #print("response status:",response.status_code)
+    if response.ok:
+        with open(filename, mode = 'w') as outfile:
+            if (len(response.content) > 100):
+                json.dump(json.loads(response.content), outfile)
+                print("retrieved telemetry match data")
+
+def retrieveAndWriteAllMatchData(filename, player_id, match_list, headers, overwrite = False, downloadTelemetry = False, telemetryLimit = 3):
     if (len(match_list) ==0):
         return
+
+    if (not overwrite and not downloadTelemetry and os.path.isfile(filename)):
+        print("filename",filename,"already exists. Aborting...")
+        return
+
     with open(filename, mode='w') as csv_file:
         # fetch all matches played in recent 14 days
 
@@ -85,15 +105,26 @@ def writeAllMatchStatsToCsv(filename, player_id, match_list):
                         "timeSurvived", "vehicleDestroys", "walkDistance", "weaponsAcquired", "winPlace", "winPoints", "winPointsDelta",
                         "teammates", "telemetryDataURL"]
 
-        writer = csv.DictWriter(csv_file, fieldnames=fieldnames)
-        writer.writeheader()
+        writer = None
+        if (not overwrite and os.path.isfile(filename)):
+            print("filename",filename,"already exists. Aborting...")
+        else:
+            writer = csv.DictWriter(csv_file, fieldnames=fieldnames)
+            writer.writeheader()
 
-        for m in match_list:
-            match_data = requestMatchStats(m["id"], player_id, headers=header, match_url=MATCH_API_REQUEST_ENDPOINT)
+        for i, m in enumerate(match_list):
+            match_data = requestMatchStats(m["id"], player_id, headers=headers, match_url=MATCH_API_REQUEST_ENDPOINT)
+            if match_data["responseStatus"] != 200:
+                print("ERROR: FAILED TO FETCH MATCH DATA")
+                continue
+
+            if downloadTelemetry and i<telemetryLimit:
+                downloadAndWriteMatchTelemetryData(PLAYERS_DATA_PATH+"telemetry/"+player_id+"_"+match_data["matchId"]+".json",
+                                                    match_data["telemetryDataURL"], headers = headers)
 
             roster_str = ""
             for teammate in match_data["rosterIdList"]:
-                roster_str+=teammate["name"]+"&"
+                roster_str+=teammate["id"]+"&"
             roster_str = roster_str[:-1]
 
             match_data = {**match_data, **match_data["matchAttributes"], **match_data["playerStats"]}
@@ -103,9 +134,54 @@ def writeAllMatchStatsToCsv(filename, player_id, match_list):
                                 "name", "responseStatus", "playerStats"]:
                 match_data.pop(unused_key)
 
-            writer.writerow(match_data)
+            if writer != None:
+                writer.writerow(match_data)
+
+def retrieveAndWritePlayerSeasonalStatsFromAPI(filename, player_id, metadata, headers, overwrite = False):
+    if (not overwrite and os.path.isfile(filename)):
+        print("filename",filename,"already exists. Aborting...")
+        return
+
+    curr_season_id = ""
+    prev_season_id = ""
+    for season in metadata["seasons"]["data"]:
+        if season["attributes"]["isCurrentSeason"]:
+            curr_season_id = season["id"]
+    prev_season_id = metadata["seasons"]["data"][-1]["id"]
+
+    req_url = "https://api.pubg.com/shards/kakao/players/"+player_id+"/seasons/"
+    req_url2 = req_url + curr_season_id
+    req_url = req_url + prev_season_id
+    current_season = {}
+    previous_season = {}
+    two_seasons_data = {}
+
+    print("retrieving seasonal stats")
+    response = requests.get(req_url, headers=headers)
+    print("response status:",response.status_code)
+    while response.status_code == 429:
+        time.sleep(6)
+        response = requests.get(req_url, headers=headers)
+        print("response status:",response.status_code)
+    if response.ok:
+        current_season = json.loads(response.content)
+
+    response = requests.get(req_url2, headers=headers)
+    print("response status:",response.status_code)
+    while response.status_code == 429:
+        time.sleep(6)
+        response = requests.get(req_url2, headers=headers)
+        print("response status:",response.status_code)
+    if response.ok:
+        previous_season = json.loads(response.content)
+
+    two_seasons_data = {"currentSeason": current_season,
+                        "previousSeason": previous_season}
+    with open(filename, mode = 'w') as outfile:
+        json.dump(two_seasons_data, outfile)
 
 def requestAndProcessPlayerDataFromAPI(url, headers):
+    global metadata
     response = requests.get(url, headers=headers)
     print("response status:",response.status_code)
     if response.ok:
@@ -113,14 +189,40 @@ def requestAndProcessPlayerDataFromAPI(url, headers):
         # get list of matches of each player
         for player_data in players_dict["data"]:
             print("retrieving",len(player_data["relationships"]["matches"]["data"]),"matches of", player_data["attributes"]["name"])
-            writeAllMatchStatsToCsv("processed_data/"+player_data["attributes"]["name"]+".csv",
+            retrieveAndWritePlayerSeasonalStatsFromAPI(PLAYERS_DATA_PATH+"seasonal_stats/"+player_data["id"]+".json",
+                                                        player_data["id"], metadata, headers)
+            retrieveAndWriteAllMatchData(PLAYERS_DATA_PATH+"match_stats/"+player_data["id"]+".csv",
                                     player_data["id"],
-                                    player_data["relationships"]["matches"]["data"])
+                                    player_data["relationships"]["matches"]["data"],
+                                    headers = headers,
+                                    overwrite = False,
+                                    downloadTelemetry = True,
+                                    telemetryLimit = 3
+                                    )
     return response.status_code
 
+
+def retrieveUpdatedMetadata():
+    with open("metadata.json") as metadata_file:
+        metadata = json.load(metadata_file)
+        last_modified = metadata["seasons"]["lastModified"]
+        # update season data every 3 days
+        if (int(round(time.time() * 1000)) - last_modified > 259200000):
+            response = requests.get("https://api.pubg.com/shards/kakao/seasons/", headers=header)
+            if response.ok:
+                metadata["seasons"] = json.loads(response.content)
+                metadata["seasons"]["lastModified"] = int(round(time.time() * 1000))
+                with open("metadata.json", "w") as metadata_outfile:
+                    json.dump(metadata, metadata_outfile)
+                    print("seasonal information updated and checked")
+        else:
+            print("seasonal information up to date")
+        return metadata
+
+
 def run_crawler():
-    for filename in os.listdir(BANNED_PLAYERS_DIRECTORY):
-        banned_players = open(BANNED_PLAYERS_DIRECTORY+"/"+filename).read().split()
+    for filename in os.listdir(PLAYERS_ID_PATH):
+        banned_players = open(PLAYERS_ID_PATH+"/"+filename).read().split()
         for i in range(0,len(banned_players),6):
             urlstring = ""
             for name in banned_players[i:min(i+6, len(banned_players))]:
@@ -130,4 +232,5 @@ def run_crawler():
             while(requestAndProcessPlayerDataFromAPI(urlstring, headers=header)==429):
                 time.sleep(6)
 
+metadata = retrieveUpdatedMetadata()
 run_crawler()
